+++ src/richie/apps/search/defaults.py
@@ -84,7 +84,7 @@ FILTERS_CONFIGURATION = [
         },
     ),
     (
-        "richie.apps.search.filter_definitions.IndexableMPTTFilterDefinition",
+        "richie.apps.search.filter_definitions.IndexableHierarchicalFilterDefinition",
         {
             "human_name": _("Subjects"),
             "is_autocompletable": True,
@@ -97,7 +97,7 @@ FILTERS_CONFIGURATION = [
         },
     ),
     (
-        "richie.apps.search.filter_definitions.IndexableMPTTFilterDefinition",
+        "richie.apps.search.filter_definitions.IndexableHierarchicalFilterDefinition",
         {
             "human_name": _("Levels"),
             "is_autocompletable": True,
@@ -110,7 +110,7 @@ FILTERS_CONFIGURATION = [
         },
     ),
     (
-        "richie.apps.search.filter_definitions.IndexableMPTTFilterDefinition",
+        "richie.apps.search.filter_definitions.IndexableHierarchicalFilterDefinition",
         {
             "human_name": _("Organizations"),
             "is_autocompletable": True,