+++ src/richie/apps/search/filter_definitions/__init__.py
@@ -8,7 +8,7 @@ from .base import NestingWrapper  # noqa
 from .courses import (  # noqa
     AvailabilityFilterDefinition,
     IndexableFilterDefinition,
-    IndexableMPTTFilterDefinition,
+    IndexableHierarchicalFilterDefinition,
     LanguagesFilterDefinition,
     StaticChoicesFilterDefinition,
 )